requirements.txt
# Competitive Research Multi-Agent System Requirements
# Python 3.8+ required

# Core dependencies
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=0.19.0

# Web scraping and HTTP requests  
aiohttp>=3.8.0
requests>=2.28.0

# Tavily AI SDK
tavily-python>=0.3.0

# CLI and output formatting
rich>=13.0.0
click>=8.0.0

# Async support
asyncio-throttle>=1.0.0

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Logging and utilities
python-json-logger>=2.0.0